@@ -2,67 +2,96 @@
   "name": "@hb42/farc",
   "displayname": "Datei-Archiv",
   "description": "Auslagern und Zurücksichern für den FI-Filer",
-  "version": "2.1.0-alpha.429",
-  "copyright": "Copyright (c) 2016-2017 Hartmut Bauer",
+  "version": "3.0.0-beta.0",
+  "copyright": "Copyright (c) 2011-2018 Hartmut Bauer",
   "author": "hb",
   "license": "MIT",
-  "private": true,
   "main": "start-electron.js",
   "scripts": {
+    "build:prep": "npm --no-git-tag-version version prerelease",
     "ng": "ng",
-    "start": "npm --no-git-tag-version version prerelease && ng serve --host=calvados --env=dev",
-    "startspk": "npm --no-git-tag-version version prerelease && ng serve --host=5.77.32.210 --env=devspk",
-    "build": "npm --no-git-tag-version version prerelease && ng build --env=prod --aot",
-    "buildspk": "npm --no-git-tag-version version prerelease && ng build --env=prodspk --aot",
-    "release": "npm version patch && ng build --env=prodsp --aot",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "electron": "cd dist && electron ."
+    "serve": "ng serve --host=calvados --configuration=devx",
+    "develx": "ng build --configuration=devx",
+    "devel": "ng build --configuration=dev",
+    "prodx": "ng build --configuration=prodx --aot",
+    "prod": "ng build --configuration=prod --aot",
+    "electron:prep": "cd dist && npm install electron-window-state",
+    "electron:pack:osx": "electron-packager dist farc --arch x64 --icon div/electron-packager/farc --overwrite --platform darwin --out ../electron-packager/ --no-prune",
+    "electron:pack:win": "electron-packager dist farc --arch x64 --icon div/electron-packager/farc --overwrite --platform win32 --out ../electron-packager/ --no-prune",
+    "electron:build:osx": "npm run electron:prep && npm run electron:pack:osx",
+    "electron:build:win": "npm run electron:prep && npm run electron:pack:win",
+    "build": "npm run build:prep && npm run prod",
+    "buildx": "npm run build:prep && npm run prodx",
+    "release": "npm --no-git-tag-version version patch && ng build --configuration=prod --aot",
+    "electronx": "npm run electron:build:osx && open ../electron-packager/farc-darwin-x64/farc.app --args dev",
+    "electron": "npm run electron:build:win && ..\\electron-packager\\farc-win32-x64\\farc.exe dev",
+    "start": "npm run build && npm run electron",
+    "startx": "npm run buildx && npm run electronx",
+    "devx": "npm run develx && npm run build:prep && npm run electronx",
+    "dev": "npm run devel && npm run build:prep && npm run electron",
+    "pack": "npm run electron:build:win && cd ../electron-packager && zip -FS -r farc-win64.zip farc-win32-x64",
+    "__rem": [
+      "build/buildx        = make /dist ",
+      "electron/electronx  = /dist to electron-package & run ",
+      "start/startx/devx   = build + electron",
+      "pack                = make win-spk-zip (electron + zip)",
+      "release             = set release version & build prod"
+    ]
+  },
+  "repository": "hb42/farc",
+  "publishConfig": {
+    "registry": "http://e077npm"
   },
   "dependencies": {
-    "@angular/animations": "^4.0.1",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "@types/mongoose": "^4.7.9",
-    "bootstrap": "^3.3.7",
+    "@angular/animations": "7.1.4",
+    "@angular/cdk": "^7.2.0",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/forms": "7.1.4",
+    "@angular/http": "7.1.4",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/router": "7.1.4",
+    "@hb42/lib-client": "^1.0.0",
+    "@hb42/lib-common": "^1.0.0",
+    "@hb42/lib-farc": "^1.0.0",
+    "@types/mongoose": "^5.3.5",
+    "bootstrap": "^4.2.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.4.1",
+    "core-js": "^2.6.1",
     "font-awesome": "^4.7.0",
-    "intl": "^1.2.5",
-    "mongoose": "^4.10.4",
-    "primeng": "^4.0.2",
-    "rxjs": "^5.1.0",
-    "semver": "^5.3.0",
-    "web-animations-js": "^2.2.5",
-    "zone.js": "^0.8.7"
+    "mongoose": "^5.4.1",
+    "popper.js": "^1.14.6",
+    "primeicons": "^1.0.0",
+    "primeng": "^7.0.3",
+    "rxjs": "^6.3.3",
+    "semver": "^5.6.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.1.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
-    "@types/jasmine": "2.5.45",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "electron": "~1.6.5",
-    "electron-window-state": "^4.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "@angular-devkit/build-angular": "^0.11.4",
+    "@angular/cli": "^7.1.4",
+    "@angular/compiler-cli": "7.1.4",
+    "@angular/language-service": "7.1.4",
+    "@types/jasmine": "^3.3.5",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^10.12.18",
+    "@types/semver": "^5.5.0",
+    "codelyzer": "^4.5.0",
+    "electron": "^4.0.0",
+    "electron-window-state": "^5.0.3",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "^3.1.4",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "node-sass": "^4.11.0",
+    "protractor": "^5.4.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "3.1.6"
   }
 }
