@@ -7,5 +7,6 @@ export * from "./roles.pipe";
 export * from "./adminview.component/adminview.component";
 export * from "./drivelist.component/drivelist.component";
 export * from "./oelist.component/oelist.component";
-export * from "./ep-list/ep-list.component";
-export * from "./config/config.component";
+export * from "./ep-list.component/ep-list.component";
+export * from "./config.component/config.component";
+export * from "./admin.guard";
