@@ -3,7 +3,9 @@
  */
 
 export * from "./config.service";
-export * from "./status";
+export * from "./main-nav.service";
+export * from "./status.component";
 export * from "./main-header.component/main-header.component";
 export * from "./user-data";
 export * from "./user-session";
+export * from "./whats.new";
