@@ -1,25 +1,24 @@
-import {
-  Pipe,
-  PipeTransform
-} from "@angular/core";
+import { Pipe, PipeTransform } from "@angular/core";
 
-import {
-  FarcSelectType,
-} from "../../../shared/ext/lib-farc";
+import { FarcSelectType, } from "@hb42/lib-farc";
 
 @Pipe({
-  name: "vormerk"
-})
+        name: "vormerk"
+      })
 export class VormerkPipe implements PipeTransform {
 
   transform(value: any, args?: any): any {
-    const sel: FarcSelectType = value;
-    switch (sel) {
-      case FarcSelectType.none        : return "";
-      case FarcSelectType.toArchive   : return "Archivieren";
-      case FarcSelectType.fromArchive : return "Zurücksichern";
-      case FarcSelectType.del         : return "Löschen";
-      default                         : return "";
+    switch (value) {
+      case FarcSelectType.none        :
+        return "";
+      case FarcSelectType.toArchive   :
+        return "Archivieren";
+      case FarcSelectType.fromArchive :
+        return "Zurücksichern";
+      case FarcSelectType.del         :
+        return "Löschen";
+      default                         :
+        return "";
     }
   }
 
