@@ -1,7 +1,7 @@
 /**
- * Created by hb on 22.04.17.
+ * electron Start-Script
  */
-
+ 
 const electron = require('electron')
 // Module to control application life.
 const app = electron.app
@@ -16,24 +16,79 @@ const windowStateKeeper = require('electron-window-state');
 const path = require('path')
 const url = require('url')
 
+// Kommandozeile
+//   - command line param[1] "dev" startet die Entwicklertools mit der Anwendung
+const development = process.argv[1] === "dev";
+
+const menuTemplate = [
+  {
+    label: 'Datei',
+    submenu: [
+      // Standard-Reload funktioniert nicht mit Angular-App
+      {label: 'Neu laden',
+        click: function(){ startApp() },
+        accelerator: process.platform !== "darwin" ? "F5" : "Cmd+R"
+	  },
+      {type: 'separator'},
+      {role: 'quit', label: "Datei-Archiv beenden"},
+    ]
+  },
+  {
+	  // default accelerators funktionieren nicht -> im Menue ausblenden
+    label: 'Fenster',
+    submenu: [
+      {role: 'minimize', label: "Minimieren", accelerator: ""},
+      {type: 'separator'},
+      {role: 'resetzoom', label: "Originalgröße", accelerator: ""},
+      {role: 'zoomin', label: "Vergrößern", accelerator: ""},
+      {role: 'zoomout', label: "Verkleinern", accelerator: ""},
+      {type: 'separator'},
+	  {label: "Entwicklertools",
+	   click: () => mainWindow.webContents.openDevTools(),
+       accelerator: process.platform !== "darwin" ? "Shift+Ctrl+I" : "Alt+Cmd+I",
+	   //visible: development,
+	  },
+    ]
+  }
+]
+
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 var mainWindow = null;
 
-var createWindow = function() {
+const startApp = function() {
+	console.log("start app");
+  // Tell Electron where to load the entry point from
+  mainWindow.loadURL(url.format({ pathname: path.join(__dirname, 'index.html'),
+                                  protocol: 'file:',
+                                  slashes: true
+                                }));
+};
 
-  ipcMain.on('asynchronous-message', function(event, arg) {
-    console.log("async msg " + arg);
-    event.sender.send('asynchronous-reply', 'pong')
-  })
+const createWindow = function() {
 
-  ipcMain.on('synchronous-message', function(event, arg) {
-    console.log("sync msg " + arg);
-    event.returnValue = 'pong'
-  })
+  // Test
+  // ipcMain.on('asynchronous-message', function(event, arg) {
+  //   console.log("async msg " + arg);
+  //   event.sender.send('asynchronous-reply', 'pong')
+  // })
+  //
+  // ipcMain.on('synchronous-message', function(event, arg) {
+  //   console.log("sync msg " + arg);
+  //   event.returnValue = 'pong'
+  // })
 
-   // Load the previous state with fallback to defaults
-  let mainWindowState = windowStateKeeper({
+  ipcMain.on("reload-app", function(event, arg) {
+    console.log("APP RELOAD " + arg);
+    startApp();
+  });
+
+  ipcMain.on("get-version", function(event, arg) {
+    event.returnValue = process.versions.electron;
+  });
+
+   // Load the previous window state with fallback to defaults
+  var mainWindowState = windowStateKeeper({
     defaultWidth: 1200,
     defaultHeight: 900
   });
@@ -43,21 +98,20 @@ var createWindow = function() {
     'x': mainWindowState.x,
     'y': mainWindowState.y,
     'width': mainWindowState.width,
-    'height': mainWindowState.height
+    'height': mainWindowState.height,
+    "icon": "favicon.ico",
+    "autoHideMenuBar": true, 
+    "webPreferences": {
+      "devTools": true
+    }
   });
 
-    // Let us register listeners on the window, so we can update the state
+  // Let us register listeners on the window, so we can update the state
   // automatically (the listeners will be removed when the window is closed)
   // and restore the maximized or full screen state
   mainWindowState.manage(mainWindow);
 
-  // session.defaultSession.allowNTLMCredentialsForDomains("*.intern");
-  // Tell Electron where to load the entry point from
-  mainWindow.loadURL(url.format({
-                                  pathname: path.join(__dirname, 'index.html'),
-                                  protocol: 'file:',
-                                  slashes: true
-                                }))
+  startApp();
 
   // Clear out the main window when the app is closed
   mainWindow.on('closed', function () {
@@ -87,13 +141,13 @@ app.on('activate', function () {
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 app.on('ready', function () {
-  createWindow();
-  // TODO nur fuer dev oeffnen
-  mainWindow.webContents.openDevTools();
-});
 
-console.log("electron app using:");
-console.log("Electron " + process.versions.electron);
-console.log("Node.js  " + process.versions.node);
-console.log("Chromium " + process.versions.chrome);
-console.log("V8       " + process.versions.v8);
+  createWindow();
+     
+  const menu = electron.Menu.buildFromTemplate(menuTemplate);
+  electron.Menu.setApplicationMenu(menu);
+
+  if (development) {
+    mainWindow.webContents.openDevTools();
+  }
+});
